class_assignments/langchain_prompts.py
from dotenv import load_dotenv
load_dotenv()
import os
from langchain_google_genai import ChatGoogleGenerativeAI

llm = ChatGoogleGenerativeAI(model="gemini-1.5-flash",
                             google_api_key=os.getenv("GOOGLE_API_KEY"),
                             ) # type:ignore
result = llm.invoke("What is the meaning of life?")