pyproject.toml
[project]
name = "agenticai-assignments"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain-core>=0.3.29",
    "langchain-google-genai>=2.0.7",
    "python-dotenv>=1.0.1",
]
